Updated a simplied version of my previous behavior manager
This new behavior manager is a more simplified version of the previous behavior manager I created in the summer. The current status is that the behavior manager is able to successfully switch between behaviors based on Lidar Sensor Data. The behavior manager has three behaviors as of today; Takeoff, Hover and Land.
These are the three that must be focused on for now. Things that must be worked on would be the bridging between behavior manager and the controller.